List Functionality, Transitions, More Comments
Added significantly more list functionality to the application, allowing user to save password that they are satisfied with. On button press it resets all text boxes and adds password to next screen

Fixed transitions more to make sense directionally and not move all over the place. Less disorientating for the user

Adding significant amounts of documentation to the python file to explain thought process and show how functions are working as well as their prupose